Refactor Program class to follow SRP
